devtools/server/tests/chrome/test_inspector-search-front.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=835896
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 835896</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  let walkerFront = null;
  let inspectorCommand = null;

  // WalkerFront and Inspector Command specific tests.  These aren't to exercise search
  // edge cases so much as to test the state the Front maintains between
  // searches.

  addAsyncTest(async function setup() {
    info("Setting up inspector and walker actors.");

    const url = document.getElementById("inspectorContent").href;

    const { commands } = await attachURL(url);
    const target = commands.targetCommand.targetFront;
    const inspector = await target.getFront("inspector");

    walkerFront = inspector.walker;
    inspectorCommand = commands.inspectorCommand;

    runNextTest();
  });

  addAsyncTest(async function testWalkerFrontDefaults() {
    info("Testing search API using WalkerFront and Inspector Command.");
    const nodes = await walkerFront.querySelectorAll(walkerFront.rootNode, "h2");
    const fronts = await nodes.items();

    const commandResult = await inspectorCommand.findNextNode("");
    ok(!commandResult, "Null result on front when searching for ''");

    let results = await inspectorCommand.findNextNode("h2");
    isDeeply(results, {
      node: fronts[0],
      resultsIndex: 0,
      resultsLength: 3,
    }, "Default options work");

    results = await inspectorCommand.findNextNode("h2", { });
    isDeeply(results, {
      node: fronts[1],
      resultsIndex: 1,
      resultsLength: 3,
    }, "Search works with empty options");

    // Clear search data to remove result state on the front
    await inspectorCommand.findNextNode("");
    runNextTest();
  });

  addAsyncTest(async function testMultipleSearches() {
    info("Testing search API using WalkerFront and Inspector Command (reverse=false)");
    const nodes = await walkerFront.querySelectorAll(walkerFront.rootNode, "h2");
    const fronts = await nodes.items();

    let results = await inspectorCommand.findNextNode("h2");
    isDeeply(results, {
      node: fronts[0],
      resultsIndex: 0,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=false)");

    results = await inspectorCommand.findNextNode("h2");
    isDeeply(results, {
      node: fronts[1],
      resultsIndex: 1,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=false)");

    results = await inspectorCommand.findNextNode("h2");
    isDeeply(results, {
      node: fronts[2],
      resultsIndex: 2,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=false)");

    results = await inspectorCommand.findNextNode("h2");
    isDeeply(results, {
      node: fronts[0],
      resultsIndex: 0,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=false)");

    // Clear search data to remove result state on the front
    await inspectorCommand.findNextNode("");
    runNextTest();
  });

  addAsyncTest(async function testMultipleSearchesReverse() {
    info("Testing search API using WalkerFront and Inspector Command (reverse=true)");
    const nodes = await walkerFront.querySelectorAll(walkerFront.rootNode, "h2");
    const fronts = await nodes.items();

    let results = await inspectorCommand.findNextNode("h2", {reverse: true});
    isDeeply(results, {
      node: fronts[2],
      resultsIndex: 2,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=true)");

    results = await inspectorCommand.findNextNode("h2", {reverse: true});
    isDeeply(results, {
      node: fronts[1],
      resultsIndex: 1,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=true)");

    results = await inspectorCommand.findNextNode("h2", {reverse: true});
    isDeeply(results, {
      node: fronts[0],
      resultsIndex: 0,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=true)");

    results = await inspectorCommand.findNextNode("h2", {reverse: true});
    isDeeply(results, {
      node: fronts[2],
      resultsIndex: 2,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=true)");

    results = await inspectorCommand.findNextNode("h2", {reverse: false});
    isDeeply(results, {
      node: fronts[0],
      resultsIndex: 0,
      resultsLength: 3,
    }, "Search works with multiple results (reverse=false)");

    // Clear search data to remove result state on the command
    await inspectorCommand.findNextNode("");
    runNextTest();
  });

  runNextTest();
};
  </script>
</head>
<body>
<a id="inspectorContent" target="_blank" href="inspector-search-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
